fix the systemd transient timer
systemd-run creates a transient timer and it inherits the same
environment variables as the systemd user environment. Since essential
env vars like PASSWORD_STORE_DIR and PASSWORD_STORE_ENABLE_EXTENSIONS
may not be present in the systemd user environemnt, we have export them
when running a transient timer so that the timer can find the `close`
command supplied by the pass-coffin extension.