fix: don't print scary messages
If a user tried to open a coffin which didn't exist, he was getting a
message saying that the coffin doesn't exist. I don't know about others
but this scared me for a second and I thought, "Wait, where is my
coffin? Did it get deleted?"

This commit tries to find `.gpg` files in a password store directory and
if it finds them, it warns the user that the password store may already
be in a decrypted state and the number of password store files that it
found. Only if it doesn't find a coffin and any `.gpg` files does
pass-coffin now print an error saying that it was unable to find a
coffin. Hopefully, nobody should this error.

This fixes #4.